get port number from .env file instead, and small changes in error handling of delete method

+++ backend/controllers/product.controller.js
@@ -12,7 +12,6 @@ export const getProducts = async (req, res) => {
     }
 }
 
-
 export const createProduct = async (req, res) => {
     const product = req.body; // user will send this data
 
@@ -30,7 +29,6 @@ export const createProduct = async (req, res) => {
         res.status(500).json({ success: false, message: "Server Error" });
     }
 }
-
 
 export const updateProduct = async (req, res) => {
     const {id} = req.params;
@@ -53,11 +51,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success:false, message:'Product not found' })
+    }
+    
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({ success: true, message: "Product deleted" })
     } catch (error) {
         console.error("Error message: ", error.message)
-        res.status(404).json({ success:"false", message: "Product not found" })
+        res.status(500).json({ success:"false", message: "Server Error" })
     }
 }
